+++ data.txt
@@ -1,2 +1,2 @@
-E|1|cs2103t|2pm
-
+D|0|resume|2022-09-01 23:59
+E|1|bday|2022-09-13 00:00
