+++ src/main/java/Todo.java
@@ -1,6 +1,4 @@
-import java.lang.annotation.Inherited;
-
-public class Todo extends Task{
+public class Todo extends Task {
 
     public static final String ENCODED_TASK_TYPE = "T";
